Fixed Bundler inline snippet to install gems by default.
The begin...rescue block protects against users who might being using
the wrong version of Bundler.

Defaulting to `true` is important as [Bundler defaults to not
installing gems](https://is.gd/ySqrjK) which can be confusing and lead
to errors when providing the script for others to use which might not
have the same setup as you.